# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/locality": {
            "post": {
                "description": "Create a new locality with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locality"
                ],
                "parameters": [
                    {
                        "description": "Locality object to be created",
                        "name": "locality",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Locality"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Locality created successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.LocalityInput"
                        }
                    }
                }
            }
        },
        "/locality/report": {
            "get": {
                "description": "Generates a report of sellers based on the provided locality ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locality"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Locality ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report of sellers by locality",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.LocalityReport"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Locality": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "locality_name": {
                    "type": "string"
                },
                "province_name": {
                    "type": "string"
                }
            }
        },
...
        "domain.LocalityInput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_province": {
                    "type": "integer"
                },
                "locality_name": {
                    "type": "string"
                }
            }
        },
        "domain.LocalityReport": {
            "type": "object",
            "properties": {
                "id_locality": {
                    "type": "integer"
                },
                "locality_name": {
                    "type": "string"
                },
                "sellers_count": {
                    "type": "integer"
                }
            }
        },
...
    }
}